# Don't allocate a byte buffer if we don't need to escape