@inconshreveable 2016-06-21
cool! i wonder if there are some additional wins we could make:

1. if no characters need to be escaped, but we need quotes (a very common case), we can copy the entire value in one call without a for loop which should allow the underlying implementation to use an optimized call like `memcpy`
1. use a switch statement instead of if with boolean ORs
1. if needQuotes and needEscape both become true, skip the remaining characters