@tg 2016-06-21
Also casting string to bytes and vice versa (`return string(e.Bytes()[start:stop])`) could be easily avoided by writing quotes only if `needsQuotes` is set and returning `e.String()` directly.