@inconshreveable 2016-06-21
@tg yeah, you're absolutely right. we should actually see if we can rewrite the formatter entirely to work entirely in bytes if possible to avoid any unnecessary conversions