@ChrisHines 2016-06-21
I don't want to put a damper on all the great work being done here, but another alternative is to use the  github.com/go-logfmt/logfmt package which has many of these ideas already incorporated. PRs on that package welcome for additional speedups.