@inconshreveable 2016-06-22
argh! I had completely forgotten that you built that fantastic package. we should definitely be adopting that and not duplicate effort here =)